Karnak.drawgraph — Method

Draw a graph g using coordinates in layout to fit in a Luxor boundingbox.

Arguments

g::AbstractGraph;
    boundingbox
    layout
    margin,
    vertexfunction
    vertexlabels
    vertexshapes
    vertexshapesizes
    vertexshaperotations
    vertexstrokecolors
    vertexstrokeweights
    vertexfillcolors
    vertexlabeltextcolors
    vertexlabelfontsizes
    vertexlabelfontfaces
    vertexlabelrotations
    vertexlabeloffsetangles
    vertexlabeloffsetdistances
    edgefunction
    edgelabels
    edgelines
    edgelist
    edgecurvature
    edgestrokecolors
    edgestrokeweights
    edgedashpatterns
    edgelabelcolors
    edgelabelrotations

Functions

vertexfunction(vertex, coordinates) ->

edgefunction(from, to) ->

edgelabels(edgenumber, edgesrc, edgedest, from, to) ->

edgeline(from, to) ->

edgelines(from, to) ->

edgestrokecolors(edgenumber, from, to) ->

vertexfillcolors(vertex) ->

vertexlabels(vertex) -> string

vertexshape(vertex) ->

vertexstrokecolors(vertex) ->

vertexstrokecolors(vertex) ->

Extended help

g

the graph to be drawn

boundingbox

the drawing fits in this BoundingBox

layout

the layout method or coordinates to be used. Examples: layout = squaregrid

layout = shell

layout = vcat( between.(O + (-W/2, H), O + (W/2, H), range(0, 1, length=N)), between.(O + (-W/2, -H), O + (W/2, -H), range(0, 1, length=N)))

layout = stress

layout = (g) -> spectral(adjacency_matrix(g), dim=2)

layout = shell ∘ adjacency_matrix

layout = (g) -> sfdp(g, Ptype=Float64, dim=2, tol=0.05, C=0.4, K=2)

layout = Shell(nlist=[6:10,])

layout = squaregrid

Refer to the NetworkLayout.jl documentation for more.

margin = 20

a margin added to the graph diagram before fitting to boundingbox

vertexfunction

a function vertexfunction(vertex, coordinates) that completely specifies the appearance of every vertex. None of the other vertex- keyword arguments will be used. Example:

vertexfunction = (v, c) -> ngon(c[v], 30, 6, 0, :fill)

edgefunction

a function edgefunction(from, to) that completely specifies the appearance of every vertex. None of the other edge- keyword arguments are used.

vertexlabels

can be function vertexlabels(vertex) -> string the text labels for each vertex. Vertex labels are not drawn by default.

vertexshapes

the shape of each vertex; can be :circle :square, or a function function vertexshape(vertex) -> Vertex center is at 0/0, vertex is vertex number, using current vertex rotation (vertextshaperotations) The function can override rotations and colors. default is ?

vertexshapesizes

the size of each vertex shape for :circle :square...

vertexshaperotations

relative to vertex locations

vertexstrokecolors

can be function: f(vertex) -> sethue("blue").

vertexstrokeweights

vertexfillcolors

can be function: f(vertex) -> sethue("blue") if both vertexfillcolors and vertexstrokecolors are functions that change Luxor's current color, fill takes priority

vertexlabeltextcolors

vertexlabelfontsizes

vertexlabelfontfaces

vertexlabelrotations

vertexlabeloffsetangles

vertexlabeldistances

edgelabels

can be function: edgelabels(edgenumber, edgesrc, edgedest, from::Point, to::Point) can be Dict can be array of strings can be a range not drawn by default, nust be specified

edgelist

list of edges

edgelines

can be function edgeline(from, to)

edgecurvature=0.0

edgestrokecolors

can be array, colorant, function: edgestrokecolors(edgenumber, from, to) edgestrokecolors = (n, from, to) -> sethue(HSB(rescale(n, 1, ne(g), 0, 360), 0.9, 0.8))

edgestrokeweights

edgedashpatterns

a dash pattern or an array of dash patterns for the edge lines: edgedashpatterns = [[10, 30], [1]]) # lines alternate between dots and lines

edgelabelcolors

the colors of the label text

edgelabelrotations

the rotation of the label text